feat(ui): migrate right rail into "You" sidebar destination
Retires the fixed 340 px right rail. Its identity readout + notification
feed are now a single content page reached via the new
SidebarNavId.PlayerInfo entry, placed above Blessings in the renamed
"OF SELF" chapter. Unread notification count rides on the sidebar item
as a badge, reusing the same mechanism as ReligionInvites.

MainLayoutCoordinator drops the rail split; content now spans from the
sidebar edge to the dialog edge, unblocking wider layouts for the
Religion Info pane. RightRailState/Event/ViewModel/Renderer are
renamed under PlayerInfo namespaces; ReligionHeaderRenderer and the
notification row code are reused verbatim. Notification feed header
adopts SubsectionLabel + Gold to match other content-page sections.
