+++ DivineAscension/GUI/State/PlayerInfoState.cs
@@ -1,11 +1,11 @@
 namespace DivineAscension.GUI.State;
 
 /// <summary>
-///     Runtime state for the right-hand rail introduced by the UI refactor.
-///     Currently tracks scroll position for the notification feed and an
+///     Runtime state for the "You" content page (former right rail).
+///     Tracks scroll position for the notification feed and an
 ///     unread-only filter toggle.
 /// </summary>
-public class RightRailState
+public class PlayerInfoState
 {
     public float ScrollY { get; set; }
     public bool ShowUnreadOnly { get; set; }
