+++ DivineAscension/assets/divineascension/lang/en.json
@@ -13,6 +13,7 @@
   "divineascension:ui.tab.religion": "Religion",
   "divineascension:ui.tab.blessings": "Blessings",
   "divineascension:ui.tab.civilization": "Civilization",
+  "divineascension:ui.tab.playerinfo": "You",
   "divineascension:ui.religion.tab.browse": "Other Orders",
   "divineascension:ui.religion.tab.info": "This Order",
   "divineascension:ui.religion.tab.activity": "Annals",
@@ -181,7 +182,7 @@
   "divineascension:ui.civilization.tab.milestones": "Chronicles",
   "divineascension:ui.sidebar.group.religion": "I. OF ORDERS",
   "divineascension:ui.sidebar.group.civilization": "II. OF REALMS",
-  "divineascension:ui.sidebar.group.personal": "III. OF BLESSINGS",
+  "divineascension:ui.sidebar.group.personal": "III. OF SELF",
   "divineascension:ui.sidebar.disabled.need_religion": "Join a religion to use this.",
   "divineascension:ui.sidebar.disabled.need_civilization": "Join a civilization to use this.",
   "divineascension:ui.sidebar.disabled.need_invite": "No invites available.",
